ai-service/requirements.txt
flask==2.3.2
flask-cors==4.0.0
pandas==2.0.3
scikit-learn==1.3.0
numpy==1.24.3
joblib==1.3.1
matplotlib==3.7.2
seaborn==0.12.2
geopy==2.3.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0 